Optimise response formatting
formatResponse is no longer called in the generation function itself. Instead, it is used by the provider. This also makes the responses consistent between what the provider generates, and what is actually displayed.
