+++ package.json
@@ -843,7 +843,7 @@
   "scripts": {
     "build": "ray build -e dist",
     "dev": "ray develop",
-    "fix-lint": "ray lint --fix",
+    "fix-lint": "ray lint --fix --relaxed",
     "lint": "ray lint",
     "publish": "npx @raycast/api@latest publish"
   }
